requirements.txt
asgiref==3.8.1
attrs==22.1.0
coverage==7.3.2
Django==5.0.3
django-cors-headers==3.13.0
djangorestframework==3.15.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.4.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.13.2
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.7.0
pytest-jest==0.3.0
pytest-metadata==2.0.2
python-dateutil==2.9.0.post0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.10.0
